Move main contract, fragment, presenter under main folder
