+++ android/app/src/main/java/com/bounswe2020group3/paperlayer/main/MainContract.kt
@@ -1,4 +1,4 @@
-package com.bounswe2020group3.paperlayer
+package com.bounswe2020group3.paperlayer.main
 
 import com.bounswe2020group3.paperlayer.base.BasePresenter
 import com.bounswe2020group3.paperlayer.base.BaseView
