+++ android/app/src/main/res/navigation/my_nav.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/mainFragment"
-        android:name="com.bounswe2020group3.paperlayer.MainFragment"
+        android:name="com.bounswe2020group3.paperlayer.main.MainFragment"
         android:label="fragment_main"
         tools:layout="@layout/fragment_main" >
         <action
